Improve menus and move files
The In-Game and Pre-Game menus have been improved in design (not quite sure of the result still).
A fade IN / OUT animation has been added between scenes.
Some files have changed location for a better understanding of the project hierarchy.

NB : The right hand and the soul are still bugged from the previous commit.